+++ server.js
@@ -666,14 +666,21 @@ app.engine('handlebars', exphbs({
     // Servers status check request
     ServerStatus() {
       let html = "";
-      if (config.casparcg){
+      if (config.casparcg) {
       config.casparcg.servers.forEach((element,i) => {
+
+        let disabledClass = ''
+        if (element.disabled) {
+          disabledClass = 'disabledServer';
+        }
+
         html += '<div class="dropdown serverbtn">';
-        html += '<i id="indicator' + i + '" class="fas fa-check-circle" style="color: 00CC00;"></i>&nbsp;' + element.name;
-        html += '  <div class="dropdown-content">';
-        html += '  <span style="opacity: 0.4; background-color: #000;">' + element.host + ':' + element.port + '</span><br>';
-        html += '  <A href="#" OnClick="spx_system(\'CHECKCONNECTIONS\');return false;">Connection check</A><br>';
-        html += '  <A href="#" OnClick="clearUsedChannels(\'' + element.name + '\');">Clear all layers</A><br>';
+        html += '   <i id="indicator' + i + '" class="fas fa-check-circle" style="color: 00CC00;"></i>&nbsp;<span class="serverIndicator ' + disabledClass + '" id="server' + element.name + '">' + element.name + '</span>';
+        html += '   <div class="dropdown-content">';
+        html += '     <span style="display: block; opacity: 0.4; background-color: #000; width:100%;">' + element.host + ':' + element.port + '</span>';
+        html += '     <A href="#" OnClick="spx_system(\'CHECKCONNECTIONS\');return false;">Connection check</A><br>';
+        html += '     <A href="#" OnClick="clearUsedChannels(\'' + element.name + '\');">Clear all layers</A><br>';
+        html += '     <A href="#" OnClick="toggleActive(\'' + element.name + '\');">Toggle on/off</A><br>';
         html += '</div>';
         html += '</div>';
       });